MVC: move all run-control/handler logic from the pane to the dock pane (#471)
ProtocolTreePane becomes a pure view that takes only view-building deps and
exposes thin view methods (select_row, enter_*_buttons, show/stop/freeze/
resume_loading, split_to_phase_controls, selected_step_path). The dock pane
is now the controller: it owns the executor, the status + logging controllers
and lifecycle handlers, the button-state machine, all executor-signal
handlers, run-control (play/pause/start), and step/phase navigation, driving
the view via those thin methods and the status model.

The pane no longer references executor or status_controller at all; run-state
(_current_row, _current_run_preview_mode, _start_pending, _wait_active) moved
to the dock pane. The status controller is now view-free.
